pyproject.toml
[project]
name = "permit-fastmcp"
version = "0.1.0"
description = "Permit.io authorization middleware for FastMCP servers."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.10.1",
    "permit>=2.7.5",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.0.0",
    "starlette>=0.47.1",
    "pyjwt>=2.8.0",
]
license = {text = "Apache-2.0"}
authors = [
    { name = "Or Weis", email = "or@permit.io" }
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
